% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-cFreeze.Rmd.
\name{freeze}
\alias{freeze}
\title{Find the Percent Historical Freeze For Any Given Month and Day}
\usage{
freeze(lon, lat, Month, Mday)
}
\arguments{
\item{lon}{GPS longitude}

\item{lat}{GPS latitude}

\item{Month}{numeric values 1:12}

\item{Mday}{numeric values 1:31 for a given month, excluding Feb-29}
}
\description{
Returns the percent of days with freezing temperatures in the last years, based on GPS coordinates and weather data from Iowa Environmental Mesonet (IEM). Note Google maps provides lat, long in that order. For this package you need to enter long, lat in that order. You must have packages apsimx and tidyverse installed.
}