% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-cFreeze.Rmd.
\name{freeze20}
\alias{freeze20}
\title{Last Day To Harvest With 20pct Historical Freeze}
\usage{
freeze20(lon, lat)
}
\arguments{
\item{lon}{GPS longitude}

\item{lat}{GPS latitude}
}
\description{
Returns the last day to harvest with up to 20pct history of freezing temperatures, based on GPS coordinates and 30-year weather data from Iowa Environmental Mesonet (IEM). For this package you need to enter long, lat in that order. Note other applications, such as Google maps, provide lat, long in the reverse order.  You must have the apsimx package installed.
}