% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-cFreeze.Rmd.
\name{nofreeze}
\alias{nofreeze}
\title{Last Day To Harvest With No Historical Freeze}
\usage{
nofreeze(lon, lat)
}
\arguments{
\item{lon}{GPS longitude}

\item{lat}{GPS latitude}
}
\description{
Returns the last day to harvest with no history of freezing temperatures, based on GPS coordinates the and last 30 years weather data from Iowa Environmental Mesonet (IEM). Note Google maps provides lat, long in that order. For this package you need to enter long, lat in that order. You must have packages apsimx and tidyverse installed.
}