% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-cFreeze.Rmd.
\name{plotdaysB0}
\alias{plotdaysB0}
\title{Plot Percent Days Below 0C}
\usage{
plotdaysB0(lon, lat)
}
\arguments{
\item{lon}{GPS longitude}

\item{lat}{GPS latitude}
}
\description{
Takes GPS coordinates, extracts last 30 years weather data from Iowa Environmental Mesonet (IEM) and returns a plot with the percent of days with minimum temperatures below 0C. Note Google maps provides lat, long in that order.  For this package you need to enter long, lat in that order. You must have packages apsimx and tidyverse installed.
}