% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-cFreeze.Rmd.
\name{freeze}
\alias{freeze}
\title{Find the Percent Historical Freeze For Any Given Month and Day}
\usage{
freeze(lon, lat, Month, Mday)
}
\arguments{
\item{lon}{GPS longitude}

\item{lat}{GPS latitude}

\item{Month}{numeric values 1:12}

\item{Mday}{numeric values 1:31 for a given month, excluding Feb-29}
}
\description{
Returns the historical percent of days with freezing temperatures for any day of the year, based on GPS coordinates and 30-year weather data from Iowa Environmental Mesonet (IEM). For this package you need to enter long, lat in that order. Note other applications, such as Google maps, provide lat, long in the reverse order.  You must have the apsimx package installed.
}