% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-cFreeze.Rmd.
\name{plotdaysB0}
\alias{plotdaysB0}
\title{Plot Percent Days Below 0C}
\usage{
plotdaysB0(lon, lat)
}
\arguments{
\item{lon}{GPS longitude}

\item{lat}{GPS latitude}
}
\description{
Returns a plot with the historical percent of days with temperatures below 0C for each day of the year, based on GPS coordinates and 30-year weather data from Iowa Environmental Mesonet (IEM). For this package you need to enter long, lat in that order. Note other applications, such as Google maps, provide lat, long in the reverse order.  You must have the apsimx package installed.
}